+++ Cargo.toml
@@ -12,7 +12,10 @@ members = [
     "examples/basic_file",
     "examples/log_ssh",
     "examples/streaming",
-    "examples/work-dispatching" 
+    "examples/work-dispatching-v1",
+    "examples/work-dispatching-v2",
+    "examples/work-dispatching-v3", 
+    "examples/work-dispatching-v4" 
 ]
 resolver = "2"
