+++ examples/work-dispatching-v1/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "work-dispatching-v1"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+env_logger = "0.8.4"
+retina-core = { path = "../../core" }
+retina-filtergen = { path = "../../filtergen" }
+retina-datatypes = { path = "../../datatypes" }
+retina-multicore = { path = "../../multicore" }
+lazy_static = "1.4.0"
+serde = { version = "1.0", features = ["derive"] }
+regex = "1.7.3"
