+++ examples/work-dispatching-v3/README.md
@@ -1,3 +1,7 @@
 # Work Dispatching
 
 An introductory example that demonstrates dispatching work to processing threads, removing subscription processing off of the critical path.
+
+Specification: 
+- Rx core agnostic routing to worker threads
+- Dedicated worker threads per callback
