+++ examples/work-dispatching-v4/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "work-dispatching"
+name = "work-dispatching-v4"
 version = "0.1.0"
 edition = "2021"
