+++ examples/work-dispatching-v4/README.md
@@ -1,0 +1,7 @@
+# Work Dispatching
+
+An introductory example that demonstrates dispatching work to processing threads, removing subscription processing off of the critical path.
+
+Specification: 
+- Rx core agnostic routing to worker threads
+- Shared worker threads across callbacks
